+++ tst/snapshots/Badge/index.test.tsx
@@ -2,15 +2,19 @@
 // SPDX-License-Identifier: Apache-2.0
 
 describe('Badge', () => {
-    it('default status', async () => {
-      await page.goto('http://host.docker.internal:9009/iframe.html?id=ui-components-badge--basic-badge');
-      const image = await page.screenshot();
-      expect(image).toMatchImageSnapshot();
-    });
-  
-    it('alart status', async () => {
-      await page.goto('http://host.docker.internal:9009/iframe.html?id=ui-components-badge--basic-badge&knob-status=alert');
-      const image = await page.screenshot();
-      expect(image).toMatchImageSnapshot();
-    });
+  it('default status', async () => {
+    await page.goto(
+      'http://host.docker.internal:9009/iframe.html?id=ui-components-badge--basic-badge&viewMode=story'
+    );
+    const image = await page.screenshot();
+    expect(image).toMatchImageSnapshot();
+  });
+
+  it('alert status', async () => {
+    await page.goto(
+      'http://host.docker.internal:9009/iframe.html?id=ui-components-badge--basic-badge&knob-status=alert&viewMode=story'
+    );
+    const image = await page.screenshot();
+    expect(image).toMatchImageSnapshot();
+  });
 });