+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.4.0] - 2020-11-06
 
 ### Fixed
 
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add base styles to ChannelList
 - Fixed createChatBubbleList unit test that was incorrectly named and not running
 - Fixed TS error inside /InfiniteList/index.tsx.
+- Fixed small visual errors in chat componnets.
+- Fixed broken snapshot tests for Badge component.
+- Fixed ModalBody component to allow custom classNames
+- Fixed PopOver components to allow custom classNames
 
 ### Added
 
@@ -27,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Attachment icon
 - Added passthrough for a react node in the subtitle of PopOverHeader
 - Add `buttonProps` passthrough for `RosterCell` and `PopOverMenu`
+- Export Chat components and utilities
 
 ### Changed
 
@@ -36,11 +41,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed ChatBubble to composed components to support redact and edit.
 - Changed RosterCell to support extra icons and menus when running late
 - Changed ChatBubble component props for simplicity.
+- Changed ChatBubbleContainer props to be more flexible.
 
 ### Removed
 
 - Removed createChatBubbleList function.
 - Removed exports and stories for unreleased chat components
+- Removed DateHeader component
 
 ## [1.3.0] - 2020-10-09